[refactor] kakaoApiClient 위치 변경
